requirements.txt
requests
discord.py
python-dotenv
deep-translator
pytest